findBasesByMD5 now throws GaveUpException - made GaveUpException a proper class - caller to findBasesByMD5 now catches IOException instead of Exception
